weather_crashes/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
cachetools==6.2.6
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.4
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
flatbuffers==25.9.23
fonttools==4.61.1
fqdn==1.5.1
fsspec==2026.1.0
geographiclib==2.1
geopy==2.4.1
greenlet==3.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ipykernel==7.1.0
ipython==8.38.0
isoduration==20.11.0
jedi==0.19.2
jh2==5.0.10
Jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.9
lark==1.3.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.0
multidict==6.7.1
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
niquests==3.17.0
notebook_shim==0.2.4
numpy==2.2.6
openmeteo_requests==1.7.5
openmeteo_sdk==1.25.0
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
postgrest==2.27.2
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
qh3==1.5.6
realtime==2.27.2
referencing==0.37.0
requests==2.32.5
requests-cache==1.2.1
retry-requests==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.1
rpds-py==0.30.0
scipy==1.15.3
seaborn==0.13.2
Send2Trash==2.1.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
stack-data==0.6.3
statsmodels==0.14.6
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
tomli==2.4.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
url-normalize==2.2.1
urllib3==2.6.3
urllib3-future==2.15.901
wassima==2.0.4
wcwidth==0.3.1
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
